+++ examples/snake.rb
@@ -4,12 +4,14 @@ require 'glimmer/data_binding/observer'
 require_relative 'snake/presenter/grid'
 
 class Snake
-  CELL_SIZE = 25
+  CELL_SIZE = 15
+  SNAKE_MOVE_DELAY = 0.1
   include Glimmer
   
   def initialize
-    @grid = Presenter::Grid.new
-    @grid.game.start
+    @game = Model::Game.new
+    @grid = Presenter::Grid.new(@game)
+    @game.start
     create_gui
     register_observers
   end
@@ -19,30 +21,45 @@ class Snake
   end
   
   def register_observers
-    @grid.game.height.times do |row|
-      @grid.game.width.times do |column|
+    @game.height.times do |row|
+      @game.width.times do |column|
         Glimmer::DataBinding::Observer.proc do |new_color|
           @cell_grid[row][column].fill = new_color
         end.observe(@grid.cells[row][column], :color)
       end
     end
-    Glimmer::LibUI.timer(0.1) do
-      @grid.game.snake.move
+    
+    Glimmer::DataBinding::Observer.proc do |game_over|
+      Glimmer::LibUI.queue_main do
+        if game_over
+          msg_box('Game Over!', "Score: #{@game.score} | High Score: #{@game.high_score}")
+          @game.start
+        end
+      end
+    end.observe(@game, :over)
+    
+    Glimmer::LibUI.timer(SNAKE_MOVE_DELAY) do
+      unless @game.over?
+        @game.snake.move
+        @main_window.title = "Glimmer Snake (Score: #{@game.score} | High Score: #{@game.high_score})"
+      end
     end
   end
   
   def create_gui
     @cell_grid = []
-    @main_window = window('Glimmer Snake', @grid.game.width * CELL_SIZE, @grid.game.height * CELL_SIZE) {
+    @main_window = window('Glimmer Snake', @game.width * CELL_SIZE, @game.height * CELL_SIZE) {
+      resizable false
+      
       vertical_box {
         padded false
         
-        @grid.game.height.times do |row|
+        @game.height.times do |row|
           @cell_grid << []
           horizontal_box {
             padded false
             
-            @grid.game.width.times do |column|
+            @game.width.times do |column|
               area {
                 @cell_grid.last << path {
                   square(0, 0, CELL_SIZE)
@@ -51,16 +68,12 @@ class Snake
                 }
                 
                 on_key_up do |area_key_event|
-                  orientation_and_key = [@grid.game.snake.head.orientation, area_key_event[:ext_key]]
+                  orientation_and_key = [@game.snake.head.orientation, area_key_event[:ext_key]]
                   case orientation_and_key
                   in [:north, :right] | [:east, :down] | [:south, :left] | [:west, :up]
-                    @grid.game.snake.turn_right
+                    @game.snake.turn_right
                   in [:north, :left] | [:west, :down] | [:south, :right] | [:east, :up]
-                    @grid.game.snake.turn_left
-                  in [*, :right]
-                    @grid.game.snake.turn_right
-                  in [*, :left]
-                    @grid.game.snake.turn_left
+                    @game.snake.turn_left
                   else
                     # No Op
                   end